update main
update main

+++ database/migrationsold/2021_03_11_115028_create_products_table.php
@@ -13,11 +13,9 @@ class CreateProductsTable extends Migration
      */
     public function up()
     {
-        Schema::create('products', function (Blueprint $table)
+        Schema::create('products', function (Blueprint $table) 
         {
-            // $table->bigIncrements('id');
-            $table->increments('id');
-
+            $table->bigIncrements('id');
             $table->string('name', 100);
             $table->string('ar_name', 100);
             $table->string('description', 400)->nullable();
